Move the content of _awaitable in mock
This was in a distinct module because it async/await, keywords
unsupported by Python 3.4.
